+++ tests/Rest/ResourceActionGrantProcessorTest.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\AuthorizationBundle\Tests\Rest;
 
+use Dbp\Relay\AuthorizationBundle\Authorization\AuthorizationService;
 use Dbp\Relay\AuthorizationBundle\Entity\ResourceActionGrant;
 use Dbp\Relay\AuthorizationBundle\Rest\ResourceActionGrantProcessor;
 use Dbp\Relay\AuthorizationBundle\Service\InternalResourceActionGrantService;
+use Dbp\Relay\AuthorizationBundle\Tests\EventSubscriber\TestGetAvailableResourceClassActionsEventSubscriber;
 use Dbp\Relay\AuthorizationBundle\TestUtils\TestEntityManager;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
 use Dbp\Relay\CoreBundle\TestUtils\DataProcessorTester;
@@ -41,6 +43,60 @@ class ResourceActionGrantProcessorTest extends AbstractResourceActionGrantContro
         $this->assertEquals($resourceActionGrant->getAuthorizationResource()->getIdentifier(), $resourceActionGrantItem->getAuthorizationResource()->getIdentifier());
         $this->assertEquals($resourceActionGrant->getAction(), $resourceActionGrantItem->getAction());
         $this->assertEquals($resourceActionGrant->getUserIdentifier(), $resourceActionGrantItem->getUserIdentifier());
+    }
+
+    public function testResourceActionGrantAddedEvent(): void
+    {
+        $manageResourceGrant = $this->addResourceAndManageGrant();
+
+        $resourceActionGrant = new ResourceActionGrant();
+        $resourceActionGrant->setAuthorizationResource($manageResourceGrant->getAuthorizationResource());
+        $resourceActionGrant->setAction(AuthorizationService::MANAGE_ACTION);
+        $resourceActionGrant->setUserIdentifier(self::CURRENT_USER_IDENTIFIER);
+        $this->resourceActionGrantProcessorTester->addItem($resourceActionGrant);
+
+        $event = $this->testResourceActionGrantAddedEventSubscriber->getEvent();
+        $this->assertEquals(self::TEST_RESOURCE_CLASS, $event->getResourceClass());
+        $this->assertEquals(self::TEST_RESOURCE_IDENTIFIER, $event->getResourceIdentifier());
+        $this->assertEquals(AuthorizationService::MANAGE_ACTION, $event->getAction());
+        $this->assertEquals(self::CURRENT_USER_IDENTIFIER, $event->getUserIdentifier());
+        $this->assertNull($event->getDynamicGroupIdentifier());
+        $this->assertNull($event->getGroupIdentifier());
+
+        $resourceActionGrant = new ResourceActionGrant();
+        $resourceActionGrant->setAuthorizationResource($manageResourceGrant->getAuthorizationResource());
+        $resourceActionGrant->setAction(TestGetAvailableResourceClassActionsEventSubscriber::READ_ACTION);
+        $resourceActionGrant->setDynamicGroupIdentifier(AuthorizationService::DYNAMIC_GROUP_IDENTIFIER_EVERYBODY);
+        $this->resourceActionGrantProcessorTester->addItem($resourceActionGrant);
+
+        $event = $this->testResourceActionGrantAddedEventSubscriber->getEvent();
+        $this->assertEquals(self::TEST_RESOURCE_CLASS, $event->getResourceClass());
+        $this->assertEquals(self::TEST_RESOURCE_IDENTIFIER, $event->getResourceIdentifier());
+        $this->assertEquals(TestGetAvailableResourceClassActionsEventSubscriber::READ_ACTION, $event->getAction());
+        $this->assertNull($event->getUserIdentifier());
+        $this->assertEquals(AuthorizationService::DYNAMIC_GROUP_IDENTIFIER_EVERYBODY, $event->getDynamicGroupIdentifier());
+        $this->assertNull($event->getGroupIdentifier());
+
+        // test with a group grant holder and a resource collection grant
+        $group = $this->testEntityManager->addGroup();
+        $this->authorizationService->addGroup($group->getIdentifier());
+        $this->addResourceAndManageGrant(
+            resourceClass: self::TEST_RESOURCE_CLASS,
+            resourceIdentifier: null);
+
+        $resourceActionGrant = new ResourceActionGrant();
+        $resourceActionGrant->setResourceClass(self::TEST_RESOURCE_CLASS);
+        $resourceActionGrant->setAction(TestGetAvailableResourceClassActionsEventSubscriber::CREATE_ACTION);
+        $resourceActionGrant->setGroup($group);
+        $this->resourceActionGrantProcessorTester->addItem($resourceActionGrant);
+
+        $event = $this->testResourceActionGrantAddedEventSubscriber->getEvent();
+        $this->assertEquals(self::TEST_RESOURCE_CLASS, $event->getResourceClass());
+        $this->assertNull($event->getResourceIdentifier());
+        $this->assertEquals(TestGetAvailableResourceClassActionsEventSubscriber::CREATE_ACTION, $event->getAction());
+        $this->assertNull($event->getUserIdentifier());
+        $this->assertNull($event->getDynamicGroupIdentifier());
+        $this->assertEquals($group->getIdentifier(), $event->getGroupIdentifier());
     }
 
     public function testCreateResourceActionGrantWithResourceClassAndIdentifierItem(): void